updated players migration - foreign key

+++ database/migrations/2019_04_05_204518_create_players_table.php
@@ -12,7 +12,7 @@ class CreatePlayersTable extends Migration
             $table->bigIncrements('id');
             $table->string('name', 100);
             $table->tinyInteger('rating');
-            $table->tinyInteger('team_id')->unsigned();
+            $table->bigInteger('team_id')->unsigned();
             // creates relationship between players and teams:
             $table->foreign('team_id')->references('id')->on('teams')->onDelete('cascade');
             $table->timestamps();
